move log models to single file
